+++ trunk/sqltoy-orm/src/main/java/org/sagacity/sqltoy/utils/CommonUtils.java
@@ -3,21 +3,11 @@
  */
 package org.sagacity.sqltoy.utils;
 
-import java.io.BufferedReader;
 import java.io.Closeable;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.URL;
 import java.util.Collection;
 import java.util.Date;
-import java.util.Enumeration;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -34,95 +24,95 @@ public class CommonUtils {
 	 */
 	private final static Logger logger = LoggerFactory.getLogger(CommonUtils.class);
 
-	/**
-	 * @TODO 读取文件存为字符串
-	 * @param file
-	 * @param charset
-	 * @return
-	 */
-	public static String readFileAsString(Object file, String charset) {
-		return inputStream2String(getFileInputStream(file), charset);
-	}
-
-	/**
-	 * @todo 获得指定路径的文件
-	 * @param file 文件路径like:classpath:xxx.xml或xxx.xml
-	 * @return
-	 */
-	public static InputStream getFileInputStream(Object file) {
-		if (file == null)
-			return null;
-		try {
-			if (file instanceof InputStream)
-				return (InputStream) file;
-			if (file instanceof File)
-				return new FileInputStream((File) file);
-
-			String realFile = (String) file;
-			if (StringUtil.isBlank(realFile))
-				return null;
-			// 文件路径
-			if (new File(realFile).exists()) {
-				return new FileInputStream(realFile);
-			}
-			if (StringUtil.indexOfIgnoreCase(realFile.trim(), "classpath:") == 0)
-				realFile = realFile.trim().substring(10).trim();
-			if (realFile.charAt(0) == '/')
-				realFile = realFile.substring(1);
-			InputStream result = Thread.currentThread().getContextClassLoader().getResourceAsStream(realFile);
-			if (result == null) {
-				try {
-					Enumeration<URL> urls = Thread.currentThread().getContextClassLoader().getResources(realFile);
-					URL url;
-					while (urls.hasMoreElements()) {
-						url = urls.nextElement();
-						result = new FileInputStream(url.getFile());
-						if (result != null)
-							break;
-					}
-				} catch (Exception e) {
-					e.printStackTrace();
-				}
-			}
-			return result;
-		} catch (FileNotFoundException fn) {
-			fn.printStackTrace();
-		}
-		return null;
-	}
-
-	/**
-	 * 转换InputStream为String
-	 *
-	 * @param is
-	 * @param encoding
-	 * @return
-	 */
-	public static String inputStream2String(InputStream is, String encoding) {
-		if (null == is) {
-			return null;
-		}
-		StringBuilder buffer = new StringBuilder();
-		BufferedReader in = null;
-		try {
-			if (StringUtil.isNotBlank(encoding)) {
-				in = new BufferedReader(new InputStreamReader(is, encoding));
-			} else {
-				in = new BufferedReader(new InputStreamReader(is));
-			}
-			String line = "";
-			while ((line = in.readLine()) != null) {
-				buffer.append(line);
-				buffer.append("\r\n");
-			}
-		} catch (Exception e) {
-			e.printStackTrace();
-			logger.error(e.getMessage());
-		} finally {
-			closeQuietly(in);
-		}
-		return buffer.toString();
-	}
+//	/**
+//	 * @TODO 读取文件存为字符串
+//	 * @param file
+//	 * @param charset
+//	 * @return
+//	 */
+//	public static String readFileAsString(Object file, String charset) {
+//		return inputStream2String(getFileInputStream(file), charset);
+//	}
+//
+//	/**
+//	 * @todo 获得指定路径的文件
+//	 * @param file 文件路径like:classpath:xxx.xml或xxx.xml
+//	 * @return
+//	 */
+//	public static InputStream getFileInputStream(Object file) {
+//		if (file == null)
+//			return null;
+//		try {
+//			if (file instanceof InputStream)
+//				return (InputStream) file;
+//			if (file instanceof File)
+//				return new FileInputStream((File) file);
+//
+//			String realFile = (String) file;
+//			if (StringUtil.isBlank(realFile))
+//				return null;
+//			// 文件路径
+//			if (new File(realFile).exists()) {
+//				return new FileInputStream(realFile);
+//			}
+//			if (StringUtil.indexOfIgnoreCase(realFile.trim(), "classpath:") == 0)
+//				realFile = realFile.trim().substring(10).trim();
+//			if (realFile.charAt(0) == '/')
+//				realFile = realFile.substring(1);
+//			InputStream result = Thread.currentThread().getContextClassLoader().getResourceAsStream(realFile);
+//			if (result == null) {
+//				try {
+//					Enumeration<URL> urls = Thread.currentThread().getContextClassLoader().getResources(realFile);
+//					URL url;
+//					while (urls.hasMoreElements()) {
+//						url = urls.nextElement();
+//						result = new FileInputStream(url.getFile());
+//						if (result != null)
+//							break;
+//					}
+//				} catch (Exception e) {
+//					e.printStackTrace();
+//				}
+//			}
+//			return result;
+//		} catch (FileNotFoundException fn) {
+//			fn.printStackTrace();
+//		}
+//		return null;
+//	}
+//
+//	/**
+//	 * 转换InputStream为String
+//	 *
+//	 * @param is
+//	 * @param encoding
+//	 * @return
+//	 */
+//	public static String inputStream2String(InputStream is, String encoding) {
+//		if (null == is) {
+//			return null;
+//		}
+//		StringBuilder buffer = new StringBuilder();
+//		BufferedReader in = null;
+//		try {
+//			if (StringUtil.isNotBlank(encoding)) {
+//				in = new BufferedReader(new InputStreamReader(is, encoding));
+//			} else {
+//				in = new BufferedReader(new InputStreamReader(is));
+//			}
+//			String line = "";
+//			while ((line = in.readLine()) != null) {
+//				buffer.append(line);
+//				buffer.append("\r\n");
+//			}
+//		} catch (Exception e) {
+//			e.printStackTrace();
+//			logger.error(e.getMessage());
+//		} finally {
+//			closeQuietly(in);
+//		}
+//		return buffer.toString();
+//	}
 
 	/**
 	 * @todo 简单逻辑判断,只支持2个逻辑,update 2017-12-4 剔除freemarker复杂逻辑判断,减少框架依赖性
